// client/models/SalesResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SalesResponse = {
    id: string;
    firstName: string;
    lastName: string;
    mobileNumber?: string;
    isActive: boolean;
    createdAt: string;
    updatedAt: string;
};